+++ src/Support/LocalFileSystem.php
@@ -10,6 +10,10 @@ class LocalFileSystem implements FileSystem
 {
     public function removeFile(string $path): bool
     {
+        if (! $this->exists($path)) {
+            throw new FileNotFound($path);
+        }
+
         return unlink($path);
     }
